Move StsDownloadButton to components
It is generic enough to be used outside the markdown component
